Remove redundant/misleading catalog_service_nodes metric
First of all, the current implementation assumes that there is only one
instance of a service running on a node. This is nowhere enforced in
consul and doesn't reflect the truth for most setups. It's completely
normal use case to run multiple service instance on the same node.
Therfore, the current metric is misleading at least, as it increases the
"total node" counter for each instance on a node.

In the end, the intended information is already available anyway, as
`consul_catalog_service_nodes{service="foo"} == count(consul_catalog_service_node_healthy{service="foo"})`