Fix and simplify collect code
The channel communication is completely useless and just makes the code
harder to follow. As an unbuffered channel is used, queryClient will
block anyway until setMetrics is called.

Furthermore, I'm very certain that the implementation is buggy and does
not export all metrics available. The reason is that the `for running`
loop will be stopped as soon as the first channel got closed, not
waiting for the other channel to be drained as well.